voice/requirements.txt
# Speech-to-text
openai>=1.0.0

# LLM extraction (Gemini -- also in ingestion/requirements.txt)
google-generativeai>=0.8.0

# Pydantic for extraction schemas
pydantic>=2.0.0

# FastAPI server (shared with ingestion)
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# FHIR models (shared with ingestion)
fhir.resources>=8.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
